--- superxmarket/Screens/Login/Extensions/LoginViewControllerConstraints.swift
@@ -1,13 +1,0 @@
-//
-//  LoginViewControllerConstraints.swift
-//  superxmarket
-//
-//  Created by Rafael Nunes Rios on 1/25/21.
-//
-
-import Foundation
-import UIKit
-
-extension LoginViewController {
-    
-}
