--- superxmarket/Supporting Files/Assets.xcassets/Icons/Contents.json
@@ -1,6 +1,0 @@
-{
-  "info" : {
-    "author" : "xcode",
-    "version" : 1
-  }
-}
