--- superxmarket/Supporting Files/Assets.xcassets/Icons/lock.imageset/Contents.json
@@ -1,22 +1,0 @@
-{
-  "images" : [
-    {
-      "filename" : "baseline_lock_white_18dp.png",
-      "idiom" : "universal",
-      "scale" : "1x"
-    },
-    {
-      "filename" : "baseline_lock_white_18dp-1.png",
-      "idiom" : "universal",
-      "scale" : "2x"
-    },
-    {
-      "idiom" : "universal",
-      "scale" : "3x"
-    }
-  ],
-  "info" : {
-    "author" : "xcode",
-    "version" : 1
-  }
-}
